// packages/client/src/types/models/checkoutLogTypes.ts
import { DeviceModel, EmployeeModel, ErrorLogModel, StudentModel } from ".";

export interface CheckoutLogModel {
  _id: any;
  device: DeviceModel;
  checkOutDate: Date;
  checkInDate?: Date;
  dueDate?: Date;
  deviceUser: StudentModel;
  teacherCheckOut: EmployeeModel;
  teacherCheckIn?: EmployeeModel;
  checkedIn: boolean;
  error?: ErrorLogModel;
  status: CheckOutStatus;
}

type CheckOutStatus = "Checked In" | "Checked Out" | "Checked In /w Error";